fix: harden the bank helpers after review
Deposits.depositAll wraps the extract-to-submit region so a synchronous
throw from the transport or the reporter refunds the bills and releases
the in-flight guard instead of losing both. Both guards are cleared on
ServerStoppedEvent like the command cooldowns.

BillDelivery no longer uses placeItemBackInInventory, which dropped the
overflow on the ground: bills are inserted through PlayerMainInvWrapper
and whatever does not fit is queued in PendingBillsData, on delivery and
on login alike. The SavedData is flushed after every add and take so a
crash between login and the next autosave cannot hand the bills out
twice. Reporter callbacks receive the player re-resolved on the server
thread rather than the instance captured at call time.

Withdrawals.validate is split out and distinguishes a malformed packet
(logged) from an empty request; the custom bundle parser accumulates in
long and rejects oversized parts with a specific message. Bills.value
returns long and PendingBillsData.add saturates.

Both helpers refuse before touching the inventory when CurrencyApi is
not initialised, so a singleplayer /deposit leaves the inventory as it
was. The Stock Ticker path delivers through BillDelivery so a full
inventory can no longer discard the withdrawn stack; its blocking join
is left for the follow-up PR.

DENOMINATIONS moves into Bills (CreateringtonCurrency keeps an alias) so
the pure helpers load without the mod's registries, and JUnit tests
cover Bills.breakdown, Withdrawals.validate and PendingBillsData.

+++ src/main/java/com/saunhardy/createringtoncurrency/client/DepositorScreen.java
@@ -413,7 +413,7 @@ public class DepositorScreen extends AbstractContainerScreen<DepositorMenu> {
         return Component.translatable("gui.createringtoncurrency.depositor." + key, args);
     }
 
-    private static String fmt(int amount) {
+    private static String fmt(long amount) {
         return NumberFormat.getInstance().format(amount);
     }
